Updated TerrainComponent to use Texture wrapper
ITexture SetData now takes a const void* instead of a void*
ITexture GetSize now returns unsigned vec2 instead of int vec2
DirectXTexture now creates an RGBA texture when calling LoadTexture(width,height) instead of pure grayscale
Fixed cast error in the window class
Added SetTexture method to the Material class
Added OnInternalChange event to the Texture class
Added Load(width,height) and SetData method to the Texture class
Fixed null exception errors in the texture class
GetSize method in the Texture class now returns unsigned vec2 instead of int vec2
Updated TerrainComponent to use a Texture wrapper to sample its terrain, Added GetHeights and SetHeights methods
Fixed all terrain components having the same material

+++ LeapEngine/GameContext/Window.cpp
@@ -49,7 +49,7 @@ void leap::Window::SetIcon(const std::string& path) const
 	Texture iconTexture{ path, true };
 	auto data = iconTexture.GetData();
 	GLFWimage images[1];
-	images[0] = GLFWimage{ iconTexture.GetSize().x, iconTexture.GetSize().y, data.data() };
+	images[0] = GLFWimage{ static_cast<int>(iconTexture.GetSize().x), static_cast<int>(iconTexture.GetSize().y), data.data() };
 
 	glfwSetWindowIcon(m_pWindow, 1, images);
 }
